requirements.txt
PyQt6-Fluent-Widgets[full]
pycryptodome
pyjwt[crypto]